use float type

+++ src/main/java/smart/dto/express/ProvinceFee.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.PositiveOrZero;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,11 +14,11 @@ public class ProvinceFee {
     @NotNull
     @PositiveOrZero
     // 首重价格
-    private Long firstFee;
+    private BigDecimal firstFee;
     @NotNull
     @PositiveOrZero
     // 续重价格
-    private Long additionalFee;
+    private BigDecimal additionalFee;
 
     public List<Long> getProvinces() {
         return provinces;
@@ -27,19 +28,19 @@ public class ProvinceFee {
         this.provinces = provinces;
     }
 
-    public Long getFirstFee() {
+    public BigDecimal getFirstFee() {
         return firstFee;
     }
 
-    public void setFirstFee(Long firstFee) {
+    public void setFirstFee(BigDecimal firstFee) {
         this.firstFee = firstFee;
     }
 
-    public Long getAdditionalFee() {
+    public BigDecimal getAdditionalFee() {
         return additionalFee;
     }
 
-    public void setAdditionalFee(Long additionalFee) {
+    public void setAdditionalFee(BigDecimal additionalFee) {
         this.additionalFee = additionalFee;
     }
 }
